requirements.txt
apturl==0.5.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
bcrypt==3.2.0
binance-futures-connector==2.1.1
black==22.3.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
cryptography==37.0.2
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
Flask==2.1.1
future==0.18.2
gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==0.4.3
netifaces==0.11.0
numpy==1.22.4
oauthlib==3.2.0
olefile==0.46
pandas==1.4.2
paramiko==2.9.3
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.20.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.27.1
SecretStorage==3.3.1
service-identity==21.1.0
six==1.16.0
ssh-import-id==5.11
systemd-python==234
ta==0.10.1
TA-Lib==0.4.24
tomli==2.0.1
ttystatus==0.38
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
virtualenv==20.14.1
wadllib==1.3.6
Werkzeug==2.1.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
